Add scroller component
